backend/requirements.txt
langchain_community==0.0.17
langchain==0.1.5
langchain_openai==0.0.5
pinecone-client==3.0.2
python-dotenv
unstructured
unstructured[md]
flask

